+++ packages/jest-emotion/src/index.js
@@ -4,6 +4,7 @@ import {
   replaceClassNames,
   type ClassNameReplacer
 } from './replace-class-names'
+import { getClassNamesFromNodes, isReactElement, isDOMElement } from './utils'
 import type { Emotion } from 'create-emotion'
 
 export { createMatchers } from './matchers'
@@ -27,30 +28,6 @@ function getNodes(node, nodes = []) {
   return nodes
 }
 
-function getClassNames(selectors, classes) {
-  return classes ? selectors.concat(classes.split(' ')) : selectors
-}
-
-function getClassNamesFromProps(selectors, props) {
-  return getClassNames(selectors, props.className || props.class)
-}
-
-function getClassNamesFromDOMElement(selectors, node) {
-  return getClassNames(selectors, node.getAttribute('class'))
-}
-
-export function getClassNamesFromNodes(nodes) {
-  return nodes.reduce(
-    (selectors, node) =>
-      isReactElement(node)
-        ? getClassNamesFromProps(selectors, node.props)
-        : isEnzymeElement(node)
-          ? getClassNamesFromProps(selectors, node.props())
-          : getClassNamesFromDOMElement(selectors, node),
-    []
-  )
-}
-
 export function getStyles(emotion: Emotion) {
   return Object.keys(emotion.caches.inserted).reduce((style, current) => {
     if (emotion.caches.inserted[current] === true) {
@@ -58,25 +35,6 @@ export function getStyles(emotion: Emotion) {
     }
     return style + emotion.caches.inserted[current]
   }, '')
-}
-
-function isReactElement(val) {
-  return val.$$typeof === Symbol.for('react.test.json')
-}
-
-function isEnzymeElement(val) {
-  return typeof val.findWhere === 'function'
-}
-
-const domElementPattern = /^((HTML|SVG)\w*)?Element$/
-
-function isDOMElement(val) {
-  return (
-    val.nodeType === 1 &&
-    val.constructor &&
-    val.constructor.name &&
-    domElementPattern.test(val.constructor.name)
-  )
 }
 
 export function createSerializer(
